refactor: move domain usecases  tests to tests folder
